requirements.txt
zhipuai
python-dotenv
pydantic
requests
Pillow
ascii-magic
rich
